Unified ShowInventory and EquipItems
basically now before you can use ShowInventory you have to call setTarget(SolShip solship) to set the SolShip whose inventory you're viewing
